refactor: reorganize transaction builder into separate files
